pyproject.toml
[project]
name = "real-human-trader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "openai-agents>=0.2.4",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
]
